+++ CHANGELOG.md
@@ -137,6 +137,7 @@
 - PR #3389 Move quantiles.hpp + group_quantiles.hpp files to legacy
 - PR #3398 Move reshape.hpp files to legacy
 - PR #3425 Strings column copy_if_else implementation
+- PR #3422 Move utilities to legacy
 
 ## Bug Fixes
 
@@ -193,6 +194,7 @@
 - PR #3413 Fix dask_cudf read_csv file-list bug
 - PR #3416 Fix memory leak in ColumnVector when pulling strings off the GPU
 - PR #3424 Fix benchmark build by adding libcudacxx to benchmark's CMakeLists.txt
+- PR #3435 Fix diff and shift for empty series
 
 # cuDF 0.10.0 (16 Oct 2019)